+++ src/components/custom-field-input-number/custom-field-input-number.jsx
@@ -48,7 +48,6 @@ export default function CustomFieldInputNumber(props) {
       placeholder={props.placeholder}
       readOnly={props.readOnly}
       required={props.required}
-      respectNativeValidity
       step={props.step}
       type="number"
     />