+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
-  - Minor: Update `CustomFieldInputNumber` to accept error and helpText props
+  - Major: All custom fields now use a single `errorText` prop to determine error state
   - Minor: Update `CustomFieldInputSingleChoice` to select choices provided via props
   - Minor: Update `CustomFieldInputSingleChoice` to allow searching and filtering on provided choices
   - Minor: Update `CustomFieldInputMultipleChoice` to select choices and deselect choices